backend/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
openai==1.54.3
pinecone-client==5.0.1
httpx==0.27.0
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.0